Prod -All

+++ src/main/java/com/ats/adminpanel/model/RawMaterial/RawMaterialDetails.java
@@ -22,7 +22,7 @@ public class RawMaterialDetails {
 	
 	private int grpId;
 	 
-	private int rmWeight;
+	private float rmWeight;
 	 
 	private int rmPackQty;
 	
@@ -100,11 +100,11 @@ public class RawMaterialDetails {
 		this.rmIcon = rmIcon;
 	}
 
-	public int getRmWeight() {
+	public float getRmWeight() {
 		return rmWeight;
 	}
 
-	public void setRmWeight(int rmWeight) {
+	public void setRmWeight(float rmWeight) {
 		this.rmWeight = rmWeight;
 	}
